Add User model, and update Teacher and Student models accordingly.
The `users` table is intended to create profiles on the site,
and to help in using one table to log users in.
Once a user logs in or registers, he will be prompted to create a profile.
The profile creation will depend on information supplied to the users table
upon registeration. The profile will for a student or either a teacher.

A user can have one of three roles:
	"admin"(Site Admin), "teacher" or a "student"